WSCOMMONS-318 Changed Part.getSize() to return a long instead of an int. Made other associated changes in Axiom. This is an internal interface, so the change in the method signature is safe.  I ran Axis2 to verify.

+++ modules/axiom-api/src/main/java/org/apache/axiom/attachments/Attachments.java
@@ -544,7 +544,7 @@ public class Attachments {
             return null;
         } else
             try {
-                int size = nextPart.getSize();
+                long size = nextPart.getSize();
                 String partContentID;
                 DataHandler dataHandler;
                 try {
