+++ modules/axiom-api/src/main/java/org/apache/axiom/attachments/impl/PartOnMemoryEnhanced.java
@@ -77,7 +77,7 @@ public class PartOnMemoryEnhanced extends AbstractPart {
         return new BAAInputStream(data, length);
     }
     
-    public int getSize() throws MessagingException {
+    public long getSize() throws MessagingException {
         return length;
     }
